lint: Prefer no arrow function parens

+++ frontend/src/api/project.ts
@@ -29,35 +29,35 @@ export function createProject(name: string, description: string): Promise<Projec
       name,
       description
     } as ProjectCreateRequestDTO)
-    .then((res) => res.data)
+    .then(res => res.data)
 }
 
 export function deleteProject(projId: string): Promise<string> {
-  return client.delete(`/api/v1/project/${projId}`).then((res) => res.data)
+  return client.delete(`/api/v1/project/${projId}`).then(res => res.data)
 }
 
 export function getAllProjects(): Promise<ProjectResponseMultipleDTO> {
-  return client.get('/api/v1/project/all').then((res) => res.data)
+  return client.get('/api/v1/project/all').then(res => res.data)
 }
 
 export function getProject(projId: string): Promise<ProjectDTO> {
-  return client.get(`/api/v1/project/${projId}`).then((res) => res.data)
+  return client.get(`/api/v1/project/${projId}`).then(res => res.data)
 }
 
 export function getProjectShares(projId: string): Promise<ProjectSharesDTO> {
-  return client.get(`/api/v1/project/${projId}/shares`).then((res) => res.data)
+  return client.get(`/api/v1/project/${projId}/shares`).then(res => res.data)
 }
 
 export function addProjectShare(projId: string, body: ProjectAddShareRequestDTO): Promise<ProjectSharesDTO> {
-  return client.post(`/api/v1/project/${projId}/shares`, body).then((res) => res.data)
+  return client.post(`/api/v1/project/${projId}/shares`, body).then(res => res.data)
 }
 
 export function deleteProjectShare(projId: string, userId: string): Promise<ProjectSharesDTO> {
-  return client.delete(`/api/v1/project/${projId}/shares/${userId}`).then((res) => res.data)
+  return client.delete(`/api/v1/project/${projId}/shares/${userId}`).then(res => res.data)
 }
 
 export function createHashlist(body: HashlistCreateRequestDTO): Promise<HashlistCreateResponseDTO> {
-  return client.post(`/api/v1/hashlist/create`, body).then((res) => res.data)
+  return client.post(`/api/v1/hashlist/create`, body).then(res => res.data)
 }
 
 export function appendToHashlist(hashlistId: string, hashes: string[]): Promise<HashlistAppendResponseDTO> {
@@ -65,57 +65,57 @@ export function appendToHashlist(hashlistId: string, hashes: string[]): Promise<
     .post(`/api/v1/hashlist/${hashlistId}/append`, {
       input_hashes: hashes
     } as HashlistAppendRequestDTO)
-    .then((res) => res.data)
+    .then(res => res.data)
 }
 
 export function createAttack(body: AttackCreateRequestDTO): Promise<AttackDTO> {
-  return client.post(`/api/v1/attack/create`, body).then((res) => res.data)
+  return client.post(`/api/v1/attack/create`, body).then(res => res.data)
 }
 
 export function deleteAttack(attackId: string): Promise<string> {
-  return client.delete(`/api/v1/attack/${attackId}`).then((res) => res.data)
+  return client.delete(`/api/v1/attack/${attackId}`).then(res => res.data)
 }
 
 export function stopAttack(attackId: string): Promise<string> {
-  return client.delete(`/api/v1/attack/${attackId}/stop`).then((res) => res.data)
+  return client.delete(`/api/v1/attack/${attackId}/stop`).then(res => res.data)
 }
 
 export function startAttack(attackId: string): Promise<AttackStartResponseDTO> {
-  return client.put(`/api/v1/attack/${attackId}/start`).then((res) => res.data)
+  return client.put(`/api/v1/attack/${attackId}/start`).then(res => res.data)
 }
 
 export function getHashlistsForProject(projId: string): Promise<HashlistResponseMultipleDTO> {
-  return client.get(`/api/v1/project/${projId}/hashlists`).then((res) => res.data)
+  return client.get(`/api/v1/project/${projId}/hashlists`).then(res => res.data)
 }
 
 export function getHashlist(hashlistId: string): Promise<HashlistDTO> {
-  return client.get(`/api/v1/hashlist/${hashlistId}`).then((res) => res.data)
+  return client.get(`/api/v1/hashlist/${hashlistId}`).then(res => res.data)
 }
 
 export function deleteHashlist(hashlistId: string): Promise<string> {
-  return client.delete(`/api/v1/hashlist/${hashlistId}`).then((res) => res.data)
+  return client.delete(`/api/v1/hashlist/${hashlistId}`).then(res => res.data)
 }
 
 export function getAttacksForHashlist(hashlistId: string): Promise<AttackMultipleDTO> {
-  return client.get(`/api/v1/hashlist/${hashlistId}/attacks`).then((res) => res.data)
+  return client.get(`/api/v1/hashlist/${hashlistId}/attacks`).then(res => res.data)
 }
 
 export function getAttacksWithJobsForHashlist(hashlistId: string, includeRuntimeData: boolean = true): Promise<AttackWithJobsMultipleDTO> {
   return client
     .get(`/api/v1/hashlist/${hashlistId}/attacks-with-jobs` + (includeRuntimeData ? '?includeRuntimeData' : ''))
-    .then((res) => res.data)
+    .then(res => res.data)
 }
 
 export function getAttacksInitialising(): Promise<AttackIDTreeMultipleDTO> {
-  return client.get('/api/v1/attack/all-initialising').then((res) => res.data)
+  return client.get('/api/v1/attack/all-initialising').then(res => res.data)
 }
 
 export function getRunningJobs(): Promise<RunningJobsForUserResponseDTO> {
-  return client.get('/api/v1/job/all-running').then((res) => res.data)
+  return client.get('/api/v1/job/all-running').then(res => res.data)
 }
 
 export function getJobCountPerUser(): Promise<RunningJobCountPerUsersDTO> {
-  return client.get('/api/v1/job/running-count-per-user').then((res) => res.data)
+  return client.get('/api/v1/job/running-count-per-user').then(res => res.data)
 }
 
 export const JobStatusCreated = 'JobStatus-Created'
